Refactors RhinoshareApp Structure
