+++ RhinoshareApp/rhinoshareapp/src/main/java/de/hhu/rhinoshareapp/domain/SecurityConfig.java
@@ -1,4 +1,4 @@
-package de.hhu.rhinoshareapp.security;
+package de.hhu.rhinoshareapp.domain;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
