+++ RhinoshareApp/rhinoshareapp/src/main/java/de/hhu/rhinoshareapp/web/MainpageController.java
@@ -1,7 +1,7 @@
 package de.hhu.rhinoshareapp.web;
 
-import de.hhu.rhinoshareapp.security.database.ServiceUser;
-import de.hhu.rhinoshareapp.security.database.ServiceUserProvider;
+import de.hhu.rhinoshareapp.domain.model.ServiceUser;
+import de.hhu.rhinoshareapp.domain.service.ServiceUserProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
