+++ RhinoshareApp/rhinoshareapp/src/main/resources/application-dev.properties
@@ -8,4 +8,4 @@ spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.hibernate.dialect=org.hibernate.dialect.H2Dialect
 spring.datasource.url=jdbc:h2:file:~/UserService
 
-server.port=80
+server.port=8080
